+++ server.js
@@ -1,7 +1,3 @@
-var multer = require("multer");
-var upload = multer({dest: "./files/"});
-
-
 const users = [
     {fullName: 'Monica Grigorovici', scsEmail: 'gmie2345@scs.ubbcluj.ro', type: 'student', password: '1234'},
     {fullName: 'Andrada Maria Gae', scsEmail: 'gaie2345@scs.ubbcluj.ro', type: 'student', password: '1234'},
@@ -27,15 +23,16 @@ const morgan = require('morgan');
 const bodyParser = require('body-parser');
 const jsSHA = require("jssha");
 
-var StudentRoutes = require('./src/routes/student-routes');
-var ProfessorRoutes = require('./src/routes/professor-routes');
-var CourseRoutes = require('./src/routes/course-routes');
-var UserRoutes = require('./src/routes/user-routes');
-var ConfigRoutes = require('./src/routes/configuration-routes');
+const StudentRoutes = require('./src/routes/student-routes');
+const ProfessorRoutes = require('./src/routes/professor-routes');
+const CourseRoutes = require('./src/routes/course-routes');
+const UserRoutes = require('./src/routes/user-routes');
+const ConfigRoutes = require('./src/routes/configuration-routes');
+const SpreadsheetRoutes = require('./src/routes/spreadsheet-routes');
 
 const app = express();
 
-app.use(bodyParser.urlencoded({extended: false}))
+app.use(bodyParser.urlencoded({extended: false}));
 app.use(bodyParser.json());
 const path = require('path');
 const port = process.env.PORT || 5000;
@@ -46,95 +43,8 @@ app.use('/course', CourseRoutes);
 app.use('/user', UserRoutes);
 app.use('/config', ConfigRoutes);
 app.use('/student', StudentRoutes);
+app.use('/spreadsheet', SpreadsheetRoutes);
 
-
-app.post('/spreadsheet', upload.single("file"), function (req, res) {
-    let students = [];
-    let disciplines = [];
-
-
-    console.log("Received file");
-    let file = req.file;
-    console.log(file.originalname);
-
-    XLSX = require('xlsx');
-    let workbook = XLSX.readFile(file.path);
-
-    let firstSheetName = workbook.SheetNames[0];
-    let addressOfCell = 'B1';
-
-    /* Get worksheet */
-    let worksheet = workbook.Sheets[firstSheetName];
-
-    /* Find desired cell */
-    let desiredCell = worksheet[addressOfCell];
-
-    /* Get the value */
-    let desiredValue = (desiredCell ? desiredCell.v : undefined);
-
-    console.log(desiredValue);
-
-    let row = 5;
-    let column = 1;
-    let cellAddress = {c: column, r: row};
-    while (worksheet[XLSX.utils.encode_cell(cellAddress)] !== undefined) {
-        let currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let formOfStudy = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let generationYear = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let section = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let discipline = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let semester = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let nrMatricol = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let lastname = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let firstname = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let email = (currentCell ? currentCell.v : undefined);
-        column++;
-
-        cellAddress = {c: column, r: row};
-        currentCell = worksheet[XLSX.utils.encode_cell(cellAddress)];
-        let group = (currentCell ? currentCell.v : undefined);
-
-        // console.log(email);
-
-        column = 1;
-        row++;
-        cellAddress = {c: column, r: row};
-    }
-
-    res.send([req.file.originalname]);
-});
 
 app.get('/check-server', (req, res) => {
     res.send({express: 'Hello From Express BACKEND!'});
